rename to feather-core
