+++ app/encoding.go
@@ -6,7 +6,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/std"
 	"github.com/cosmos/cosmos-sdk/x/auth/tx"
 
-	"github.com/terra-money/feather-base/app/params"
+	"github.com/terra-money/feather-core/app/params"
 )
 
 // makeEncodingConfig creates an EncodingConfig for an amino based test configuration.
