+++ cmd/feather-cored/cmd/genaccounts.go
