+++ docs/static/openapi.yml
@@ -3,7 +3,7 @@ swagger: '2.0'
 info:
   title: HTTP API Console
   version: ''
-  description: 'feather-base blockchain'
+  description: 'feather-core blockchain'
 paths:
   /terra/alliances:
     get:
