+++ transformations.yml
@@ -18,7 +18,7 @@ inputs:
 transformations:
     - name: project path
       type: replace
-      pattern: github.com/terra-money/feather-base
+      pattern: github.com/terra-money/feather-core
       replacement: "{{ .package_name }}"
       files: 
         - "*/**.go"
@@ -26,7 +26,7 @@ transformations:
         - "go.sum"
     - name: chain name
       type: replace
-      pattern: "feather-base"
+      pattern: "feather-core"
       replacement: "{{ .name }}"
       files: 
         - "*/**.go"
@@ -50,7 +50,7 @@ transformations:
       files: ["README.md"]
     - name: rename cmd
       type: rename
-      pattern: "feather-based"
+      pattern: "feather-cored"
       replacement: "{{ .name }}d" # The text to replace. You may use go templates and perform arbitrary replacements.
       files: 
         - "*/**"
